將 MapUI.Block 抽出成獨立的 ImageLS class
