+++ src/main/java/us/dontcareabout/gameRoom/client/mine/ui/component/MapUI.java
@@ -9,23 +9,22 @@ import us.dontcareabout.gameRoom.client.mine.ui.BoardView;
 import us.dontcareabout.gameRoom.client.mine.vo.GameInfo;
 import us.dontcareabout.gameRoom.client.mine.vo.XY;
 import us.dontcareabout.gxt.client.draw.LImageSprite;
-import us.dontcareabout.gxt.client.draw.LayerSprite;
 import us.dontcareabout.gxt.client.draw.layout.HorizontalLayoutLayer;
 import us.dontcareabout.gxt.client.draw.layout.VerticalLayoutLayer;
 
 public class MapUI extends VerticalLayoutLayer {
-	private final Block[][] map;
+	private final ImageLS[][] map;
 
 	public MapUI(int width, int height) {
 		setMargins(1);
-		map = new Block[height][width];
+		map = new ImageLS[height][width];
 
 		for (int y = 0; y < height; y++) {
 			HorizontalLayoutLayer hll = new HorizontalLayoutLayer();
 			hll.setMargins(1);
 			addChild(hll, BoardView.BLOCK_SIZE);
 			for (int x = 0; x < width; x++) {
-				map[y][x] = new Block(new LImageSprite(ImageRS.I.unknown()));
+				map[y][x] = new ImageLS(new LImageSprite(ImageRS.I.unknown()));
 				XY xy = new XY(x, y);	//lambda 裡頭得是常數
 				map[y][x].addSpriteSelectionHandler(e -> GM.move(BoardView.ID, xy));
 				hll.addChild(map[y][x], BoardView.BLOCK_SIZE);
@@ -38,28 +37,8 @@ public class MapUI extends VerticalLayoutLayer {
 
 		for (int y = 0; y < info.getHeight(); y++) {
 			for (int x = 0; x < info.getWidth(); x++) {
-				map[y][x].refresh(value[x][y]);
+				map[y][x].refresh(mapping(value[x][y]));
 			}
-		}
-	}
-
-	//Refactory GF：先有 HasSize 的 LSprite，然後弄個 SimpleLayerSprite 不然煩死人
-	private class Block extends LayerSprite {
-		private final LImageSprite sprite;
-
-		Block(LImageSprite sprite) {
-			this.sprite = sprite;
-			add(sprite);
-		}
-
-		void refresh(int value) {
-			sprite.setResource(mapping(value));
-		}
-
-		@Override
-		protected void adjustMember() {
-			sprite.setWidth(getWidth());
-			sprite.setHeight(getHeight());
 		}
 	}
